chore: refactor packages
- rename packages to more appropriate one
