+++ store/src/main/java/com/zextras/mailbox/account/usecase/FolderUtil.java
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: AGPL-3.0-only
 
-package com.zextras.mailbox.domain.usecase;
+package com.zextras.mailbox.account.usecase;
 
 import com.zimbra.cs.mailbox.Folder;
 import com.zimbra.cs.mailbox.Mailbox.FolderNode;
