+++ store/src/main/java/com/zextras/mailbox/account/usecase/UseCaseDI.java
@@ -1,7 +1,12 @@
-package com.zextras.mailbox.domain.usecase;
+// SPDX-FileCopyrightText: 2023 Zextras <https://www.zextras.com>
+//
+// SPDX-License-Identifier: AGPL-3.0-only
+
+package com.zextras.mailbox.account.usecase;
 
 import com.google.inject.AbstractModule;
 import com.google.inject.Provides;
+import com.zextras.mailbox.acl.AclService;
 import com.zimbra.common.service.ServiceException;
 import com.zimbra.common.util.Log;
 import com.zimbra.common.util.ZimbraLog;
