+++ store/src/main/java/com/zimbra/cs/service/admin/AdminService.java
@@ -5,8 +5,8 @@
 
 package com.zimbra.cs.service.admin;
 
-import com.zextras.mailbox.domain.usecase.AclService;
-import com.zextras.mailbox.domain.usecase.DeleteUserUseCase;
+import com.zextras.mailbox.account.usecase.DeleteUserUseCase;
+import com.zextras.mailbox.acl.AclService;
 import com.zimbra.common.service.ServiceException;
 import com.zimbra.common.soap.AdminConstants;
 import com.zimbra.common.soap.Element;
