+++ store/src/test/java/com/zextras/mailbox/domain/usecase/DeleteUserUseCaseTest.java
@@ -3,6 +3,8 @@ package com.zextras.mailbox.domain.usecase;
 import static org.junit.jupiter.api.Assertions.*;
 import static org.mockito.Mockito.*;
 
+import com.zextras.mailbox.account.usecase.DeleteUserUseCase;
+import com.zextras.mailbox.acl.AclService;
 import com.zimbra.common.account.ZAttrProvisioning;
 import com.zimbra.common.util.Log;
 import com.zimbra.cs.account.*;
